Moving files around + combining DFs
Files are more organised. Code to combine and initially visualise the data for every day in the set is written as well as some graphs. The code runs on the files posted in our teams chat which aren't in the repository. The code would need to be adjusted to work on files on your local machine.